[DF-9277] Fixed review issues
Former-commit-id: 529abf3c1e03cf1abdfde22d03e4c04be0ce93a7

+++ Tools/AssetCacheClient/Classes/main.cpp
@@ -108,5 +108,5 @@ int main(int argc, char* argv[])
     }
 
     ReleaseDAVA();
-    return cacheClient.GetExitCode();
+    return static_cast<int>(cacheClient.GetExitCode());
 }
